added an enemy that is long ranged
same script to follow me but add a few more components to keep distance

+++ War-Of-Shapes/Assets/Scripts/healthBar.cs
@@ -9,7 +9,7 @@ public class healthBar : MonoBehaviour
     public Gradient gradient;
     public Image fill;
 
-    public void maxHealth(int health)
+    public void maxHealth(float health)
     {
         sli.value = health;
         sli.maxValue = health;
